RangeSlider: Don't update position only if mouse/touch is grabbed
A mouseClick will cause the RangeSlider handle to move to the click
position and we should expect the same for a touch press.

We update the position of the handle in handleRelease if keepMouse/
TouchGrab is set to true. However, we don't grab the touch on a
touch press in case eg.: a flickable wants to steal the event.
So we can remove the check as we shouldn't update the position
only if grabbed. This makes the behavior similar to the slider
too.

Fixes: QTBUG-112945
Change-Id: Id2cf99416a52b5e42989a4adda1532e3ac550a93
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit f6fb9252c6af824ee4544b34a039c20872c5b4ed)